requirements.txt
Flask==0.11.1
Flask-API==0.6.6.post1
python-logstash==0.4.6
Flask-Bootstrap==3.3.6.0
Flask-SQLAlchemy==2.1
flask-iniconfig==0.0.9
requests
pyOpenSSL